datatypes::compute::kernels

Module arity

Kernels for operating on [PrimitiveArray]s

Functions

binary

Allies a binary infallable function to two [PrimitiveArray]s, producing a new [PrimitiveArray]

binary_mut

Applies a binary and infallible function to values in two arrays, replacing the values in the first array in place.

try_binary

Applies the provided fallible binary operation across a and b.

try_binary_mut

Applies the provided fallible binary operation across a and b by mutating the mutable [PrimitiveArray] a with the results.

try_unary

See [PrimitiveArray::try_unary]

try_unary_mut

See [PrimitiveArray::try_unary_mut]

unary

See [PrimitiveArray::unary]

unary_mut

See [PrimitiveArray::unary_mut]