datatypes::compute::kernels

Module bitwise

Module contains bitwise operations on arrays

Functions

bitwise_and

Perform left & right operation on two arrays. If either left or right value is null then the result is also null.

bitwise_and_not

Perform left & !right operation on two arrays. If either left or right value is null then the result is also null.

bitwise_and_scalar

Perform bitwise and every value in an array with the scalar. If any value in the array is null then the result is also null.

bitwise_not

Perform !array operation on array. If array value is null then the result is also null.

bitwise_or

Perform left | right operation on two arrays. If either left or right value is null then the result is also null.

bitwise_or_scalar

Perform bitwise or every value in an array with the scalar. If any value in the array is null then the result is also null.

bitwise_shift_left

Perform bitwise left << right operation on two arrays. If either left or right value is null then the result is also null.

bitwise_shift_left_scalar

Perform bitwise left << right every value in an array with the scalar. If any value in the array is null then the result is also null.

bitwise_shift_right

Perform bitwise left >> right operation on two arrays. If either left or right value is null then the result is also null.

bitwise_shift_right_scalar

Perform bitwise left >> right every value in an array with the scalar. If any value in the array is null then the result is also null.

bitwise_xor

Perform left ^ right operation on two arrays. If either left or right value is null then the result is also null.

bitwise_xor_scalar

Perform bitwise xor every value in an array with the scalar. If any value in the array is null then the result is also null.