datatypes::compute::kernels

Module cast_utils

Parser implementations for converting strings to Arrow types

Used by the CSV and JSON readers to convert strings to Arrow types

Structs

IntervalParseConfig

Config to parse interval strings

Enums

IntervalUnit

Represents the units of an interval, with each variant corresponding to a bit in the interval’s bitfield representation

Traits

Parser

Specialized parsing implementations to convert strings to Arrow types.

Functions

parse_decimal

Parse the string format decimal value to i128/i256 format and checking the precision and scale. The result value can’t be out of bounds.

parse_interval_day_time

Parse human-readable interval string to Arrow [IntervalDayTimeType]

parse_interval_month_day_nano

Parse human-readable interval string to Arrow [IntervalMonthDayNanoType]

parse_interval_month_day_nano_config

Parse human-readable interval string to Arrow [IntervalMonthDayNanoType]

parse_interval_year_month

Parse human-readable interval string to Arrow [IntervalYearMonthType]

string_to_datetime

Accepts a string and parses it relative to the provided timezone

string_to_time_nanoseconds

Accepts a string in ISO8601 standard format and some variants and converts it to nanoseconds since midnight.

string_to_timestamp_nanos

Accepts a string in RFC3339 / ISO8601 standard format and some variants and converts it to a nanosecond precision timestamp.

Type Aliases

MonthDayNano

A tuple representing (months, days, nanoseconds) in an interval