datatypes::compute::kernels::cast_utils

Function string_to_time_nanoseconds

pub fn string_to_time_nanoseconds(s: &str) -> Result<i64, ArrowError>

Accepts a string in ISO8601 standard format and some variants and converts it to nanoseconds since midnight.

Examples of accepted inputs:

09:26:56.123 AM

23:59:59

6:00 pm

Internally, this function uses the chrono library for the time parsing

Timezone / Offset Handling

This function does not support parsing strings with a timezone or offset specified, as it considers only time since midnight.