datatypes::compute::kernels::cmp

Function lt

pub fn lt(lhs: &dyn Datum, rhs: &dyn Datum) -> Result<BooleanArray, ArrowError>

Perform left < right operation on two [Datum].

Comparing null values on either side will yield a null in the corresponding slot of the resulting [BooleanArray].

For floating values like f32 and f64, this comparison produces an ordering in accordance to the totalOrder predicate as defined in the IEEE 754 (2008 revision) floating point standard. Note that totalOrder treats positive and negative zeros as different. If it is necessary to treat them as equal, please normalize zeros before calling this kernel. See f32::total_cmp and f64::total_cmp.

Nested types, such as lists, are not supported as the null semantics are not well-defined. For comparisons involving nested types see [crate::ord::make_comparator]