datatypes::compute::kernels

Module comparison

Comparison kernels for Arrays.

Traits

StringArrayType

A trait for Arrow String Arrays, currently three types are supported:

Functions

contains

Perform SQL CONTAINS(left, right)

ends_with

Perform SQL ENDSWITH(left, right)

ilike

Perform SQL left ILIKE right

in_list

Checks if a [GenericListArray] contains a value in the [PrimitiveArray]

in_list_utf8

Checks if a [GenericListArray] contains a value in the [GenericStringArray]

like

Perform SQL left LIKE right

nilike

Perform SQL left NOT ILIKE right

nlike

Perform SQL left NOT LIKE right

regexp_is_match

Return BooleanArray indicating which strings in an array match an array of regular expressions.

regexp_is_match_scalar

Return BooleanArray indicating which strings in an array match a single regular expression.

regexp_is_match_utf8Deprecated

Perform SQL array ~ regex_array operation on [StringArray] / [LargeStringArray]. If regex_array element has an empty value, the corresponding result value is always true.

regexp_is_match_utf8_scalarDeprecated

Perform SQL array ~ regex_array operation on [StringArray] / [LargeStringArray] and a scalar.

starts_with

Perform SQL STARTSWITH(left, right)