Trait datatypes::data_type::DataType

pub trait DataType: Debug + Send + Sync {
    // Required methods
    fn name(&self) -> &str;
    fn logical_type_id(&self) -> LogicalTypeId;
    fn default_value(&self) -> Value;
    fn as_arrow_type(&self) -> ArrowDataType;
    fn create_mutable_vector(&self, capacity: usize) -> Box<dyn MutableVector>;
    fn is_timestamp_compatible(&self) -> bool;
}

Data type abstraction.

Required Methods

fn name(&self) -> &str

Name of this data type.

fn logical_type_id(&self) -> LogicalTypeId

Returns id of the Logical data type.

fn default_value(&self) -> Value

Returns the default value of this type.

fn as_arrow_type(&self) -> ArrowDataType

Convert this type as [arrow::datatypes::DataType].

fn create_mutable_vector(&self, capacity: usize) -> Box<dyn MutableVector>

Creates a mutable vector with given capacity of this type.

fn is_timestamp_compatible(&self) -> bool

Returns true if the data type is compatible with timestamp type so we can use it as a timestamp.