Trait datatypes::scalars::ScalarVector

pub trait ScalarVector: Vector + Send + Sync + Sized + 'staticwhere
    for<'a> Self::OwnedItem: Scalar<RefType<'a> = Self::RefItem<'a>>,{
    type OwnedItem: Scalar<VectorType = Self>;
    type RefItem<'a>: ScalarRef<'a, ScalarType = Self::OwnedItem>
       where Self: 'a;
    type Iter<'a>: Iterator<Item = Option<Self::RefItem<'a>>>
       where Self: 'a;
    type Builder: ScalarVectorBuilder<VectorType = Self>;

    // Required methods
    fn get_data(&self, idx: usize) -> Option<Self::RefItem<'_>>;
    fn iter_data(&self) -> Self::Iter<'_>;

    // Provided methods
    fn from_slice(data: &[Self::RefItem<'_>]) -> Self { ... }
    fn from_iterator<'a>(it: impl Iterator<Item = Self::RefItem<'a>>) -> Self { ... }
    fn from_owned_iterator(
        it: impl Iterator<Item = Option<Self::OwnedItem>>
    ) -> Self { ... }
    fn from_vec<I: Into<Self::OwnedItem>>(values: Vec<I>) -> Self { ... }
}

A sub trait of Vector to add scalar operation support.

Required Associated Types

type OwnedItem: Scalar<VectorType = Self>

type RefItem<'a>: ScalarRef<'a, ScalarType = Self::OwnedItem> where Self: 'a

The reference item of this vector.

type Iter<'a>: Iterator<Item = Option<Self::RefItem<'a>>> where Self: 'a

Iterator type of this vector.

type Builder: ScalarVectorBuilder<VectorType = Self>

Builder type to build this vector.

Required Methods

fn get_data(&self, idx: usize) -> Option<Self::RefItem<'_>>

Returns the reference to an element at given position.

Note: get() has bad performance, avoid call this function inside loop.

Panics

Panics if idx >= self.len().

fn iter_data(&self) -> Self::Iter<'_>

Returns iterator of current vector.

Provided Methods

fn from_slice(data: &[Self::RefItem<'_>]) -> Self

fn from_iterator<'a>(it: impl Iterator<Item = Self::RefItem<'a>>) -> Self

fn from_owned_iterator( it: impl Iterator<Item = Option<Self::OwnedItem>> ) -> Self

fn from_vec<I: Into<Self::OwnedItem>>(values: Vec<I>) -> Self