datatypes::schema

Struct ColumnSchema

pub struct ColumnSchema {
    pub name: String,
    pub data_type: ConcreteDataType,
    is_nullable: bool,
    is_time_index: bool,
    default_constraint: Option<ColumnDefaultConstraint>,
    metadata: Metadata,
}

Schema of a column, used as an immutable struct.

Fields
name: Stringdata_type: ConcreteDataTypeis_nullable: boolis_time_index: booldefault_constraint: Option<ColumnDefaultConstraint>metadata: Metadata
Implementations

impl ColumnSchema

pub fn new<T: Into<String>>( name: T, data_type: ConcreteDataType, is_nullable: bool, ) -> ColumnSchema

pub fn is_time_index(&self) -> bool

pub fn is_nullable(&self) -> bool

pub fn default_constraint(&self) -> Option<&ColumnDefaultConstraint>

pub fn is_default_impure(&self) -> bool

Check if the default constraint is a impure function.

pub fn metadata(&self) -> &Metadata

pub fn mut_metadata(&mut self) -> &mut Metadata

pub fn column_comment(&self) -> Option<&String>

Retrieve the column comment

pub fn with_time_index(self, is_time_index: bool) -> Self

pub fn set_inverted_index(&mut self, value: bool)

Set the inverted index for the column. Similar to with_inverted_index but don’t take the ownership.

pub fn with_inverted_index(self, value: bool) -> Self

Set the inverted index for the column. Similar to set_inverted_index but take the ownership and return a owned value.

pub fn is_inverted_indexed(&self) -> bool

pub fn is_fulltext_indexed(&self) -> bool

pub fn is_skipping_indexed(&self) -> bool

pub fn has_inverted_index_key(&self) -> bool

pub fn with_default_constraint( self, default_constraint: Option<ColumnDefaultConstraint>, ) -> Result<Self>

Set default constraint.

If a default constraint exists for the column, this method will validate it against the column’s data type and nullability.

pub fn with_nullable_set(self) -> Self

Set the nullablity to true of the column. Similar to set_nullable but take the ownership and return a owned value.

pub fn set_nullable(&mut self)

Set the nullability to true of the column. Similar to with_nullable_set but don’t take the ownership

pub fn set_time_index(&mut self)

Set the is_time_index to true of the column. Similar to with_time_index but don’t take the ownership.

pub fn with_metadata(self, metadata: Metadata) -> Self

Creates a new ColumnSchema with given metadata.

pub fn create_default_vector( &self, num_rows: usize, ) -> Result<Option<VectorRef>>

Creates a vector with default value for this column.

If the column is NOT NULL but doesn’t has DEFAULT value supplied, returns Ok(None).

pub fn create_default_vector_for_padding(&self, num_rows: usize) -> VectorRef

Creates a vector for padding.

This method always returns a vector since it uses DataType::default_value to fill the vector. Callers should only use the created vector for padding and never read its content.

pub fn create_default(&self) -> Result<Option<Value>>

Creates a default value for this column.

If the column is NOT NULL but doesn’t has DEFAULT value supplied, returns Ok(None).

pub fn create_impure_default(&self) -> Result<Option<Value>>

Creates an impure default value for this column, only if it have a impure default constraint. Otherwise, returns Ok(None).

pub fn fulltext_options(&self) -> Result<Option<FulltextOptions>>

Retrieves the fulltext options for the column.

pub fn with_fulltext_options(self, options: FulltextOptions) -> Result<Self>

pub fn set_fulltext_options(&mut self, options: &FulltextOptions) -> Result<()>

pub fn skipping_index_options(&self) -> Result<Option<SkippingIndexOptions>>

Retrieves the skipping index options for the column.

pub fn with_skipping_options( self, options: SkippingIndexOptions, ) -> Result<Self>

pub fn set_skipping_options( &mut self, options: &SkippingIndexOptions, ) -> Result<()>

pub fn unset_skipping_options(&mut self) -> Result<()>