Enum datatypes::type_id::LogicalTypeId

pub enum LogicalTypeId {

    Null,
    Boolean,
    Int8,
    Int16,
    Int32,
    Int64,
    UInt8,
    UInt16,
    UInt32,
    UInt64,
    Float32,
    Float64,
    Decimal128,
    String,
    Binary,
    Date,
    DateTime,
    TimestampSecond,
    TimestampMillisecond,
    TimestampMicrosecond,
    TimestampNanosecond,
    TimeSecond,
    TimeMillisecond,
    TimeMicrosecond,
    TimeNanosecond,
    DurationSecond,
    DurationMillisecond,
    DurationMicrosecond,
    DurationNanosecond,
    IntervalYearMonth,
    IntervalDayTime,
    IntervalMonthDayNano,
    List,
    Dictionary,

}

Unique identifier for logical data type.

Variants

Null

Boolean

Int8

Int16

Int32

Int64

UInt8

UInt16

UInt32

UInt64

Float32

Float64

Decimal128

String

Binary

Date

Date representing the elapsed time since UNIX epoch (1970-01-01) in days (32 bits).

DateTime

Datetime representing the elapsed time since UNIX epoch (1970-01-01) in seconds/milliseconds/microseconds/nanoseconds, determined by precision.

TimestampSecond

TimestampMillisecond

TimestampMicrosecond

TimestampNanosecond

TimeSecond

A 64-bit time representing the elapsed time since midnight in the unit of TimeUnit.

TimeMillisecond

TimeMicrosecond

TimeNanosecond

DurationSecond

A 64-bit duration representing the elapsed time in either seconds, milliseconds, microseconds or nanoseconds.

DurationMillisecond

DurationMicrosecond

DurationNanosecond

IntervalYearMonth

A 32-bit interval representing the elapsed time in months.

IntervalDayTime

A 64-bit interval representing the elapsed time in days and milliseconds.

IntervalMonthDayNano

A 128-bit interval representing the elapsed time in months, days and nanoseconds.

List

Dictionary