Module datatypes::types

Re-exports

pub use cast::cast;

pub use cast::cast_with_opt;

Modules

binary_type 🔒

boolean_type 🔒

cast

date_type 🔒

datetime_type 🔒

decimal_type 🔒

dictionary_type 🔒

duration_type 🔒

interval_type 🔒

json_type 🔒

list_type 🔒

null_type 🔒

primitive_type 🔒

string_type 🔒

time_type 🔒

TimeType represents the elapsed time since midnight in the unit of TimeUnit.

timestamp_type 🔒

Structs

BinaryType

BooleanType

DateTimeType

Data type for [DateTime].

DateType

Data type for Date (YYYY-MM-DD).

Decimal128Type

Decimal type with precision and scale information.

DictionaryType

Used to represent the Dictionary datatype.

DurationMicrosecondType

DurationMillisecondType

DurationNanosecondType

DurationSecondType

Float32Type

Float64Type

Int8Type

Int16Type

Int32Type

Int64Type

IntervalDayTimeType

IntervalMonthDayNanoType

IntervalYearMonthType

JsonType

JsonType is a data type for JSON data. It is stored as binary data of jsonb format. It utilizes current binary value and vector implementation.

ListType

Used to represent the List datatype.

NullType

OrdPrimitive

A new type for WrapperType, complement the Ord feature for it. Wrapping non ordered primitive types like f32 and f64 in OrdPrimitive can make them be used in places that require Ord. For example, in Median UDAFs.

StringType

TimeMicrosecondType

TimeMillisecondType

TimeNanosecondType

TimeSecondType

TimestampMicrosecondType

TimestampMillisecondType

TimestampNanosecondType

TimestampSecondType

UInt8Type

UInt16Type

UInt32Type

UInt64Type

Enums

DurationType

IntervalType

The “calendar” interval is a type of time interval that does not have a precise duration without taking into account a specific base timestamp.

TimeType

TimestampType

Constants

JSON_TYPE_NAME

Traits

LogicalPrimitiveType

Trait bridging the logical primitive type with [ArrowPrimitiveType].

WrapperType

Represents the wrapper type that wraps a native type using the newtype pattern, such as Date is a wrapper type for the underlying native type i32.