datatypes::types::primitive_type

Trait WrapperType

pub trait WrapperType:
    Copy
    + Send
    + Sync
    + Debug
    + for<'a> Scalar<RefType<'a> = Self>
    + PartialEq
    + Into<Value>
    + Into<ValueRef<'static>>
    + Serialize
    + Into<Value> {
    type LogicalType: LogicalPrimitiveType<Wrapper = Self, Native = Self::Native>;
    type Native: ArrowNativeType;

    // Required methods
    fn from_native(value: Self::Native) -> Self;
    fn into_native(self) -> Self::Native;
}

Represents the wrapper type that wraps a native type using the newtype pattern, such as Date is a wrapper type for the underlying native type i32.

Required Associated Types

type LogicalType: LogicalPrimitiveType<Wrapper = Self, Native = Self::Native>

Logical primitive type that this wrapper type belongs to.

type Native: ArrowNativeType

The underlying native type.

Required Methods

fn from_native(value: Self::Native) -> Self

Convert native type into this wrapper type.

fn into_native(self) -> Self::Native

Convert this wrapper type into native type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.