Module datatypes::value

Macros

define_data_type_func 🔒

impl_as_for_value_ref 🔒

impl_ord_for_value_like 🔒

impl_try_as_primitive 🔒

impl_try_from_value 🔒

impl_value_from 🔒

impl_value_ref_from 🔒

Structs

DictionaryValue

Dictionary value.

ListValue

List value.

OrderedFloat

A wrapper around floats providing implementations of Eq, Ord, and Hash.

Enums

ListValueRef

Reference to a ListValue.

Value

Value holds a single arbitrary value of any DataType.

ValueRef

Reference to Value.

Traits

TryAsPrimitive

Functions

duration_to_scalar_value

Cast the 64-bit duration into the arrow ScalarValue with time unit.

new_item_field 🔒

scalar_value_to_interval

Convert [ScalarValue] to Interval.

scalar_value_to_timestamp

Convert [ScalarValue] to Timestamp. If it’s ScalarValue::Utf8, try to parse it with the given timezone. Return None if given scalar value cannot be converted to a valid timestamp.

time_to_scalar_value

Cast the 64-bit elapsed time into the arrow ScalarValue by time unit.

timestamp_to_scalar_value

to_null_scalar_value

Type Aliases

OrderedF32

OrderedF64