datatypes::value

Enum Value

pub enum Value {

    Null,
    Boolean(bool),
    UInt8(u8),
    UInt16(u16),
    UInt32(u32),
    UInt64(u64),
    Int8(i8),
    Int16(i16),
    Int32(i32),
    Int64(i64),
    Float32(OrderedF32),
    Float64(OrderedF64),
    Decimal128(Decimal128),
    String(StringBytes),
    Binary(Bytes),
    Date(Date),
    Timestamp(Timestamp),
    Time(Time),
    Duration(Duration),
    IntervalYearMonth(IntervalYearMonth),
    IntervalDayTime(IntervalDayTime),
    IntervalMonthDayNano(IntervalMonthDayNano),
    List(ListValue),
    Struct(StructValue),
    Json(Box<JsonValue>),

}

Value holds a single arbitrary value of any DataType.

Comparison between values with different types (expect Null) is not allowed.

Variants

Null

Boolean(bool)

UInt8(u8)

UInt16(u16)

UInt32(u32)

UInt64(u64)

Int8(i8)

Int16(i16)

Int32(i32)

Int64(i64)

Float32(OrderedF32)

Float64(OrderedF64)

Decimal128(Decimal128)

String(StringBytes)

Binary(Bytes)

Date(Date)

Timestamp(Timestamp)

Time(Time)

Duration(Duration)

IntervalYearMonth(IntervalYearMonth)

IntervalDayTime(IntervalDayTime)

IntervalMonthDayNano(IntervalMonthDayNano)

List(ListValue)

Struct(StructValue)

Json(Box<JsonValue>)

Implementations

impl Value

pub fn data_type(&self) -> ConcreteDataType

Returns data type of the value.

Panics

Panics if the data type is not supported.

pub fn is_null(&self) -> bool

Returns true if this is a null value.

pub fn as_list(&self) -> Result<Option<&ListValue>>

Cast itself to ListValue.

pub fn as_struct(&self) -> Result<Option<&StructValue>>

pub fn as_value_ref(&self) -> ValueRef<'_>

Cast itself to ValueRef.

pub fn as_timestamp(&self) -> Option<Timestamp>

Cast Value to timestamp. Return None if value is not a valid timestamp data type.

pub fn as_string(&self) -> Option<String>

Cast Value to utf8 String. Return None if value is not a valid string data type.

pub fn as_date(&self) -> Option<Date>

Cast Value to Date. Return None if value is not a valid date data type.

pub fn as_time(&self) -> Option<Time>

Cast Value to [Time]. Return None if value is not a valid time data type.

pub fn as_interval_year_month(&self) -> Option<IntervalYearMonth>

Cast Value to [IntervalYearMonth]. Return None if value is not a valid interval year month data type.

pub fn as_interval_day_time(&self) -> Option<IntervalDayTime>

Cast Value to [IntervalDayTime]. Return None if value is not a valid interval day time data type.

pub fn as_interval_month_day_nano(&self) -> Option<IntervalMonthDayNano>

Cast Value to [IntervalMonthDayNano]. Return None if value is not a valid interval month day nano data type.

pub fn as_i64(&self) -> Option<i64>

Cast Value to i64. Return None if value is not a valid int64 data type.

pub fn as_u64(&self) -> Option<u64>

Cast Value to u64. Return None if value is not a valid uint64 data type.

pub fn as_f64_lossy(&self) -> Option<f64>

Cast Value to f64. Return None if it’s not castable;

pub fn as_duration(&self) -> Option<Duration>

Cast Value to [Duration]. Return None if value is not a valid duration data type.

pub fn as_bool(&self) -> Option<bool>

Cast value to Boolean. Return None if value is not a boolean type.

pub fn into_json_inner(self) -> Option<Value>

Extract the inner JSON value from a JSON type.

pub fn logical_type_id(&self) -> LogicalTypeId

Returns the logical type of the value.

pub fn try_to_scalar_value( &self, output_type: &ConcreteDataType, ) -> Result<ScalarValue>

Convert the value into [ScalarValue] according to the output_type.

pub fn try_negative(&self) -> Option<Self>

Apply - unary op if possible