Module datatypes::schema

Re-exports

pub use crate::schema::constraint::ColumnDefaultConstraint;

Modules

column_schema 🔒

constraint

raw 🔒

Structs

ColumnSchema

Schema of a column, used as an immutable struct.

FieldsAndIndices 🔒

FulltextOptions

Fulltext options for a column.

RawSchema

Struct used to serialize and deserialize Schema.

Schema

A common schema, should be immutable.

SchemaBuilder

Enums

FulltextAnalyzer

Fulltext analyzer.

Constants

COMMENT_KEY

FULLTEXT_KEY

Key used to store fulltext options in arrow field’s metadata.

TIME_INDEX_KEY

Key used to store whether the column is time index in arrow field’s metadata.

TYPE_KEY

Key used to store actual column type in field metadata.

VERSION_KEY

Key used to store version number of the schema in metadata.

Functions

collect_fields 🔒

try_parse_version 🔒

validate_timestamp_index 🔒

Type Aliases

Metadata

SchemaRef