datatypes
Module schema

Re-exports

pub use crate::schema::constraint::ColumnDefaultConstraint;

Modules

column_schema 🔒

constraint

raw 🔒

Structs

ColumnSchema

Schema of a column, used as an immutable struct.

FieldsAndIndices 🔒

FulltextOptions

Fulltext options for a column.

RawSchema

Struct used to serialize and deserialize Schema.

Schema

A common schema, should be immutable.

SchemaBuilder

SkippingIndexOptions

Skipping options for a column.

Enums

FulltextAnalyzer

Fulltext analyzer.

SkippingIndexType

Skip index types.

Constants

COLUMN_FULLTEXT_CHANGE_OPT_KEY_ENABLE

Keys used in fulltext options

COLUMN_FULLTEXT_OPT_KEY_ANALYZER

COLUMN_FULLTEXT_OPT_KEY_CASE_SENSITIVE

COLUMN_SKIPPING_INDEX_OPT_KEY_GRANULARITY

Keys used in SKIPPING index options

COLUMN_SKIPPING_INDEX_OPT_KEY_TYPE

COMMENT_KEY

FULLTEXT_KEY

Key used to store fulltext options in arrow field’s metadata.

INVERTED_INDEX_KEY

Key used to store whether the column has inverted index in arrow field’s metadata.

SKIPPING_INDEX_KEY

Key used to store skip options in arrow field’s metadata.

TIME_INDEX_KEY

Key used to store whether the column is time index in arrow field’s metadata.

TYPE_KEY

Key used to store actual column type in field metadata.

VERSION_KEY

Key used to store version number of the schema in metadata.

Functions

collect_fields 🔒

try_parse_version 🔒

validate_timestamp_index 🔒

Type Aliases

Metadata

SchemaRef