Module datatypes::schema

Re-exports

pub use crate::schema::constraint::ColumnDefaultConstraint;

Modules

column_schema 🔒

constraint

raw 🔒

Structs

ColumnSchema

Schema of a column, used as an immutable struct.

FieldsAndIndices 🔒

RawSchema

Struct used to serialize and deserialize Schema.

Schema

A common schema, should be immutable.

SchemaBuilder

Constants

COMMENT_KEY

TIME_INDEX_KEY

Key used to store whether the column is time index in arrow field’s metadata.

VERSION_KEY

Key used to store version number of the schema in metadata.

Functions

collect_fields 🔒

try_parse_version 🔒

validate_timestamp_index 🔒

Type Aliases

Metadata

SchemaRef