datatypes::schema

Module column_schema

Structs

ColumnSchema

Schema of a column, used as an immutable struct.

FulltextOptions

Fulltext options for a column.

SkippingIndexOptions

Skipping options for a column.

Enums

ColumnExtType

Column extended type set in column schema’s metadata.

FulltextAnalyzer

Fulltext analyzer.

FulltextBackend

The backend of the fulltext index.

SkippingIndexType

Skip index types.

Constants

COLUMN_FULLTEXT_CHANGE_OPT_KEY_ENABLE

Keys used in fulltext options

COLUMN_FULLTEXT_OPT_KEY_ANALYZER

COLUMN_FULLTEXT_OPT_KEY_BACKEND

COLUMN_FULLTEXT_OPT_KEY_CASE_SENSITIVE

COLUMN_SKIPPING_INDEX_OPT_KEY_GRANULARITY

Keys used in SKIPPING index options

COLUMN_SKIPPING_INDEX_OPT_KEY_TYPE

COMMENT_KEY

DEFAULT_CONSTRAINT_KEY 🔒

Key used to store default constraint in arrow field’s metadata.

DEFAULT_GRANULARITY

FULLTEXT_KEY

Key used to store fulltext options in arrow field’s metadata.

INVERTED_INDEX_KEY

Key used to store whether the column has inverted index in arrow field’s metadata.

SKIPPING_INDEX_KEY

Key used to store skip options in arrow field’s metadata.

TIME_INDEX_KEY

Key used to store whether the column is time index in arrow field’s metadata.

Type Aliases

Metadata