Module datatypes::types

Modules

binary_type 🔒

boolean_type 🔒

date_type 🔒

datetime_type 🔒

dictionary_type 🔒

list_type 🔒

null_type 🔒

primitive_type 🔒

string_type 🔒

timestamp_type 🔒

Structs

BinaryType

BooleanType

DateTimeType

Data type for DateTime.

DateType

Data type for Date (YYYY-MM-DD).

DictionaryType

Used to represent the Dictionary datatype.

Float32Type

Float64Type

Int8Type

Int16Type

Int32Type

Int64Type

ListType

Used to represent the List datatype.

NullType

OrdPrimitive

A new type for WrapperType, complement the Ord feature for it. Wrapping non ordered primitive types like f32 and f64 in OrdPrimitive can make them be used in places that require Ord. For example, in Median or Percentile UDAFs.

StringType

TimestampMicrosecondType

TimestampMillisecondType

TimestampNanosecondType

TimestampSecondType

UInt8Type

UInt16Type

UInt32Type

UInt64Type

Enums

TimestampType

Traits

LogicalPrimitiveType

Trait bridging the logical primitive type with [ArrowPrimitiveType].

NativeType

Data types that can be used as arrow’s native type.

WrapperType

Represents the wrapper type that wraps a native type using the newtype pattern, such as Date is a wrapper type for the underlying native type i32.