datatypes::types

Module primitive_type

Macros

define_logical_primitive_type 🔒

define_non_timestamp_primitive 🔒

impl_wrapper 🔒

Structs

Float32Type

Float64Type

Int8Type

Int16Type

Int32Type

Int64Type

OrdPrimitive

A new type for WrapperType, complement the Ord feature for it.

UInt8Type

UInt16Type

UInt32Type

UInt64Type

Traits

LogicalPrimitiveType

Trait bridging the logical primitive type with [ArrowPrimitiveType].

WrapperType

Represents the wrapper type that wraps a native type using the newtype pattern, such as Date is a wrapper type for the underlying native type i32.