datatypes::vectors

Trait Vector

pub trait Vector:
    Send
    + Sync
    + Serializable
    + Debug
    + VectorOp {

    // Required methods
    fn data_type(&self) -> ConcreteDataType;
    fn vector_type_name(&self) -> String;
    fn as_any(&self) -> &dyn Any;
    fn len(&self) -> usize;
    fn to_arrow_array(&self) -> ArrayRef;
    fn to_boxed_arrow_array(&self) -> Box<dyn Array>;
    fn validity(&self) -> Validity;
    fn memory_size(&self) -> usize;
    fn null_count(&self) -> usize;
    fn is_null(&self, row: usize) -> bool;
    fn slice(&self, offset: usize, length: usize) -> VectorRef;
    fn get(&self, index: usize) -> Value;
    fn get_ref(&self, index: usize) -> ValueRef<'_>;

    // Provided methods
    fn is_empty(&self) -> bool { ... }
    fn is_const(&self) -> bool { ... }
    fn only_null(&self) -> bool { ... }
    fn try_get(&self, index: usize) -> Result<Value> { ... }

}

Vector of data values.

Required Methods

fn data_type(&self) -> ConcreteDataType

Returns the data type of the vector.

This may require heap allocation.

fn vector_type_name(&self) -> String

fn as_any(&self) -> &dyn Any

Returns the vector as Any so that it can be downcast to a specific implementation.

fn len(&self) -> usize

Returns number of elements in the vector.

fn to_arrow_array(&self) -> ArrayRef

Convert this vector to a new arrow [ArrayRef].

fn to_boxed_arrow_array(&self) -> Box<dyn Array>

Convert this vector to a new boxed arrow [Array].

fn validity(&self) -> Validity

Returns the validity of the Array.

fn memory_size(&self) -> usize

Returns the memory size of vector.

fn null_count(&self) -> usize

The number of null slots on this Vector.

Implementation

This is O(1).

fn is_null(&self, row: usize) -> bool

Returns whether row is null.

fn slice(&self, offset: usize, length: usize) -> VectorRef

Slices the Vector, returning a new VectorRef.

Panics

This function panics if offset + length > self.len().

fn get(&self, index: usize) -> Value

Returns the clone of value at index.

Panics

Panic if index is out of bound.

fn get_ref(&self, index: usize) -> ValueRef<'_>

Returns the reference of value at index.

Panics

Panic if index is out of bound.

Provided Methods

fn is_empty(&self) -> bool

Returns whether the vector is empty.

fn is_const(&self) -> bool

Returns true when it’s a ConstantColumn

fn only_null(&self) -> bool

If the vector only contains NULL.

fn try_get(&self, index: usize) -> Result<Value>

Returns the clone of value at index or error if index is out of bound.